% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly.R
\docType{data}
\name{bitly}
\alias{bitly}
\title{Data on browser and OS characteristics for requests made on
\code{USA.gov}.}
\description{
These URLs are created whenever anyone shortens a \code{.gov} or \code{.mil}
URL using \href{https://bitly.com/}{bitly}. The dta consists of a named
list of variables including \code{tz} for timezone and \code{a} agent.
}
\references{
\href{https://github.com/usagov/1.USA.gov-Data}{USA.gov Data}
}
\keyword{data}
