% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ourConvolve}
\alias{ourConvolve}
\title{Convolve two numeric vectors and return the resuting vector.}
\usage{
ourConvolve(a, b)
}
\arguments{
\item{a}{the first vector}

\item{b}{the second vector}
}
\value{
a new vector that the convolution
}
\description{
The convolution of two vectors \eqn{a} of length \eqn{n_1} and
and \eqn{b} of length \eqn{n_2} is of length \eqn{n_1 + n_2 - 1}.
}
\examples{
base::set.seed(1)
ourConvolve(a = rnorm(2), b = rnorm(3))

}
