% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTzByOS.R
\name{plotTzByOS}
\alias{plotTzByOS}
\title{Create a ggplot barplot of timezone by operating system}
\usage{
plotTzByOS(data, color = c("blue", "#E69F00"))
}
\arguments{
\item{data}{the browser agent data as a list containing variables \code{tz} for timezone and
\code{a} for agent}

\item{color}{colors to use for bars, default \code{blue}}
}
\description{
Generate a bar plot of browser agent data by operating system windows or nonwindows.
}
\examples{
data(bitly)
plotTzByOS(bitly)
}
\keyword{dplot}
