% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat290.ass2.R
\docType{package}
\name{stat290.ass2}
\alias{stat290.ass2}
\title{Stat 290 Simple Test Package}
\description{
\code{stat290.ass2} is a simple ("hello world") package for
stat290.
}
\details{
The package provides two functions:
\itemize{
\item \code{\link[=plotTzByOS]{plotTzByOS}} which uses \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} to produce a plot object
\item \code{\link[=ourConvolve]{ourConvolve}} which uses the \code{Rcpp} package to link R code to a C routine.
}

The source code of the package further demonstrates:
\itemize{
\item \href{https://rmarkdown.rstudio.com/}{R markdown} syntax in documenting functions, including references to functions and data within the same package and from other packages using the \href{https://cran.r-project.org/package=roxygen2/vignettes/rd-formatting.html}{\code{roxygen2} package}
\item Including a data set \code{\link{bitly}} as part of the package
\item Importing functions from other packages using \code{importFrom} directives
\item Use of \code{\link[utils:globalVariables]{utils::globalVariables()}} to get around warnings produced when checking the package, something that happens a lot when using \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}, for instance.
}
}
\seealso{
\link[=ourConvolve]{ourConvolve}

\link[=plotTzByOS]{plotTzByOS}
}
